AT3-483 fix many clang lint warnings in utils

+++ src/ska/pst/common/utils/src/Timer.cpp
@@ -3,18 +3,18 @@
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  * this list of conditions and the following disclaimer.
- * 
+ *
  * 2. Redistributions in binary form must reproduce the above copyright
  * notice, this list of conditions and the following disclaimer in the
  * documentation and/or other materials provided with the distribution.
- * 
+ *
  * 3. Neither the name of the copyright holder nor the names of its
  * contributors may be used to endorse or promote products derived from this
  * software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -59,13 +59,13 @@ auto ska::pst::common::Timer::get_elapsed_microseconds() -> double
   gettimeofday(&timestamp, nullptr);
   const time_t seconds = timestamp.tv_sec - start_epoch.tv_sec;
   const suseconds_t usec =  timestamp.tv_usec - start_epoch.tv_usec;
-  return double((seconds * microseconds_per_second) + usec);
+  return ((static_cast<double>(seconds) * microseconds_per_second) + static_cast<double>(usec));
 }
 
 auto ska::pst::common::Timer::get_elapsed_milliseconds() -> int
 {
   static constexpr double microseconds_per_millisecond = 1000;
-  return int(rint(get_elapsed_microseconds()/microseconds_per_millisecond));
+  return static_cast<int>(rint(get_elapsed_microseconds()/microseconds_per_millisecond));
 }
 
 void ska::pst::common::Timer::print_rates(uint64_t bytes)
@@ -77,7 +77,7 @@ void ska::pst::common::Timer::print_rates(uint64_t bytes)
   double gbytes_per_second = 0;
   if (elapsed_microseconds > 0)
   {
-    bytes_per_second = double(bytes) / (elapsed_microseconds / microseconds_per_second);
+    bytes_per_second = static_cast<double>(bytes) / (elapsed_microseconds / microseconds_per_second);
     gbytes_per_second = bytes_per_second / bytes_per_gigabyte;
   }
